requirements.txt
websocket-server>=0.6<1
websocket-client>1.3<2